requirements.txt
discord.py[voice]
google-generativeai
yt-dlp
python-dotenv
spotipy
PyNaCl
